DataScientestGroupProjectContributions/requirements.txt
absl-py==2.0.0
gensim==4.3.2
imbalanced-learn==0.11.0
imblearn==0.0
ipython==8.16.1
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.2
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
nltk==3.8.1
numpy==1.25.2
pandas==2.1.1
pyspellchecker==0.7.2
regex==2023.10.3
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.12.2
shap==0.44.0
sklearn==0.0.post10
streamlit==1.29.0
tqdm==4.66.1
wordcloud==1.9.2
xgboost==2.0.1
